Specify array types in doctests
On Windows, it appears that the default integer type used by NumPy is
`int32` if not otherwise specified. As these tests are based on using
`int64`, just explicitly set this type instead of allowing it to be
picked by default. This should work fine on Unix while extending this
behavior to Windows as well.