Flatten `indices` default dtype check
Use a ternary expression to fill in the `dtype` to use in `indices`, if
it is not already specified. Should note that `int` on Unix and Windows
translate to different things. On Unix, `int` is the same as `int64`. On
Windows, `int` is the same as `int32`. As such, the default type seems
to be platform dependent. However, this feels less that ideal. So we may
want to fix this to specific precision in the future. We should also add
some tests specifically for the unspecified `dtype` case.