Switch all doctest arrays from `int64` to `uint64`
It appears that on Windows the dtype will be displayed as part of the
NumPy array representation if it is `int32` on Python 3, but not Python
2. However, it will display the dtype as part of the NumPy array
representation on both Python 2 and 3 if it is `int64`. This contrasts
with Unix where a NumPy `int64` array does not have its dtype included
in the representation. To avoid this annoying issue, just pick a type
that will always show up in NumPy's array representation regardless of
OS or Python version. In particular, we pick `uint64`, which already
works for all of the arrays we have created. This way the doctests
should run without issues everywhere.